+++ src/main/java/com/fastcampus/projectboardamdin/dto/properties/ProjectProperties.java
@@ -1,17 +1,25 @@
 package com.fastcampus.projectboardamdin.dto.properties;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 @ConfigurationProperties("project")
 @Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class ProjectProperties {
 
     private Board board;
 
     @Getter
-    public class Board {
+    @Setter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class Board {
         private String url;
     }
 }
