+++ src/test/java/com/fastcampus/projectboardamdin/service/ArticleManagementServiceTest.java
@@ -1,12 +1,12 @@
 package com.fastcampus.projectboardamdin.service;
 
-import com.fastcampus.projectboardamdin.domain.constant.RoleType;
-import com.fastcampus.projectboardamdin.dto.ArticleCommentDto;
 import com.fastcampus.projectboardamdin.dto.ArticleDto;
 import com.fastcampus.projectboardamdin.dto.UserAccountDto;
 import com.fastcampus.projectboardamdin.dto.properties.ProjectProperties;
-import com.fastcampus.projectboardamdin.dto.response.ArticleCommentClientResponse;
+import com.fastcampus.projectboardamdin.dto.response.ArticleClientResponse;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -21,7 +21,6 @@ import org.springframework.test.web.client.MockRestServiceServer;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.client.match.MockRestRequestMatchers.method;
@@ -31,7 +30,7 @@ import static org.springframework.test.web.client.response.MockRestResponseCreat
 @DisplayName("비즈니스 로직 - 게시글 관리")
 class ArticleManagementServiceTest {
 
-    //@Disabled("실제 API 호출 결과를 관찰용이므로 평상시엔 비활성화")
+    @Disabled("실제 API 호출 결과를 관찰용이므로 평상시엔 비활성화")
     @DisplayName("실제 API 호출 테스트")
     @SpringBootTest
     @Nested
@@ -60,117 +59,123 @@ class ArticleManagementServiceTest {
 
     }
 
-    @DisplayName("API mocking 테스트")
+    @DisplayName("API Mocking Test")
     @EnableConfigurationProperties(ProjectProperties.class)
     @AutoConfigureWebClient(registerRestTemplate = true)
-    @RestClientTest(ArticleCommentManagementService.class)
+    @RestClientTest(ArticleManagementService.class)
     @Nested
     class RestTemplateTest {
 
-        private final ArticleCommentManagementService sut;
+        private final ArticleManagementService sut;
         private final ProjectProperties projectProperties;
         private final MockRestServiceServer server;
         private final ObjectMapper mapper;
 
         @Autowired
-        public RestTemplateTest(ArticleCommentManagementService sut, ProjectProperties projectProperties, MockRestServiceServer server, ObjectMapper mapper) {
+        public RestTemplateTest(ArticleManagementService sut, ProjectProperties projectProperties, MockRestServiceServer server, ObjectMapper mapper) {
             this.sut = sut;
             this.projectProperties = projectProperties;
             this.server = server;
             this.mapper = mapper;
         }
 
-        @DisplayName("댓글 목록 API을 호출하면, 댓글들을 가져온다.")
+        @DisplayName("게시글 목록 API를 호출하면, 게시글들을 가져온다.")
         @Test
-        void givenNothing_whenCallingCommentsApi_thenReturnsCommentList() throws Exception {
-            // Given
-            ArticleCommentDto expectedComment = createArticleCommentDto("댓글");
-            ArticleCommentClientResponse expectedResponse = ArticleCommentClientResponse.of(List.of(expectedComment));
+        void givenNothing_whenCallingArticlesApi_thenReturnArticleList() throws JsonProcessingException {
 
+            // given
+            ArticleDto expectedArticle = createArticleDto("제목", "내용");
+            ArticleClientResponse expectedResponse = ArticleClientResponse.of(List.of(expectedArticle));
             server
-                    .expect(requestTo(projectProperties.getBoard().getUrl() + "/api/articleComments?size=10000"))
+                    .expect(requestTo(projectProperties.getBoard().getUrl() + "/api/articles?size=10000"))
                     .andRespond(withSuccess(
-                            mapper.writeValueAsString(expectedResponse),
+                            mapper.writeValueAsString(expectedResponse), // JSON 변환
                             MediaType.APPLICATION_JSON
                     ));
 
+            // when
+            List<ArticleDto> result = sut.getArticles();
 
-            // When
-            List<ArticleCommentDto> result = sut.getArticleComments();
-
-            // Then
-            assertThat(result).first()
-                    .hasFieldOrPropertyWithValue("id", expectedComment.getId())
-                    .hasFieldOrPropertyWithValue("content", expectedComment.getContent())
-                    .hasFieldOrPropertyWithValue("userAccount.nickname", expectedComment.getUserAccount().getNickname());
-
-            server.verify();
-        }
-
-        @DisplayName("댓글 ID와 함께 댓글 API을 호출하면, 댓글을 가져온다.")
-        @Test
-        void givenCommentId_whenCallingCommentApi_thenReturnsComment() throws Exception {
-
-            // Given
-            Long articleCommentId = 1L;
-            ArticleCommentDto expectedComment = createArticleCommentDto("댓글");
-            server
-                    .expect(requestTo(projectProperties.getBoard().getUrl() + "/api/articleComments/" + articleCommentId))
-                    .andRespond(withSuccess(
-                            mapper.writeValueAsString(expectedComment),
-                            MediaType.APPLICATION_JSON
-                    ));
-
-            // When
-            ArticleCommentDto result = sut.getArticleComment(articleCommentId);
-
-            // Then
+            // then
             assertThat(result)
-                    .hasFieldOrPropertyWithValue("id", expectedComment.getId())
-                    .hasFieldOrPropertyWithValue("content", expectedComment.getContent())
-                    .hasFieldOrPropertyWithValue("userAccount.nickname", expectedComment.getUserAccount().getNickname());
+                    .first()
+                    .hasFieldOrPropertyWithValue("id", expectedArticle.getId())
+                    .hasFieldOrPropertyWithValue("title", expectedArticle.getTitle())
+                    .hasFieldOrPropertyWithValue("content", expectedArticle.getContent())
+                    .hasFieldOrPropertyWithValue("userAccount.nickname", expectedArticle.getUserAccount().getNickname());
+
+            // 위에서 작성한 통신이 실행되었는지 검증
             server.verify();
         }
 
-        @DisplayName("댓글 ID와 함께 댓글 삭제 API을 호출하면, 댓글을 삭제한다.")
+        @DisplayName("게시글 API를 호출하면, 게시글을 가져온다.")
         @Test
-        void givenCommentId_whenCallingDeleteCommentApi_thenDeletesComment() throws Exception {
-            // Given
-            Long articleCommentId = 1L;
+        void givenNothing_whenCallingArticleApi_thenReturnArticle() throws JsonProcessingException {
+
+            // given
+            Long articleId = 1L;
+            ArticleDto expectedArticle = createArticleDto("게시판", "글");
             server
-                    .expect(requestTo(projectProperties.getBoard().getUrl() + "/api/articleComments/" + articleCommentId))
+                    .expect(requestTo(projectProperties.getBoard().getUrl() + "/api/articles/" + articleId))
+                    .andRespond(withSuccess(
+                            mapper.writeValueAsString(expectedArticle),
+                            MediaType.APPLICATION_JSON
+                    ));
+
+            // when
+            ArticleDto result = sut.getArticle(articleId);
+
+            // then
+            assertThat(result)
+                    .hasFieldOrPropertyWithValue("id", expectedArticle.getId())
+                    .hasFieldOrPropertyWithValue("title", expectedArticle.getTitle())
+                    .hasFieldOrPropertyWithValue("content", expectedArticle.getContent())
+                    .hasFieldOrPropertyWithValue("userAccount.nickname", expectedArticle.getUserAccount().getNickname());
+
+            server.verify();
+        }
+
+        @DisplayName("게시글 ID와 함께 게시글 삭제 API를 호출하면, 게시글을 삭제한다.")
+        @Test
+        void givenArticleId_whenCallingDeleteArticleApi_thenDeleteArticle() throws JsonProcessingException {
+
+            // given
+            Long articleId = 1L;
+
+            server
+                    .expect(requestTo(projectProperties.getBoard().getUrl() + "/api/articles/" + articleId))
                     .andExpect(method(HttpMethod.DELETE))
                     .andRespond(withSuccess());
 
-            // When
-            sut.deleteArticleComment(articleCommentId);
+            // when
+            sut.deleteArticle(articleId);
 
-            // Then
+            // then
             server.verify();
+        }
+
+        private ArticleDto createArticleDto(String title, String content) {
+            return ArticleDto.of(
+                    1L,
+                    createUserAccountDto(),
+                    title,
+                    content,
+                    null,
+                    LocalDateTime.now(),
+                    "Uno",
+                    LocalDateTime.now(),
+                    "Uno"
+            );
+        }
+
+        private UserAccountDto createUserAccountDto() {
+            return UserAccountDto.of(
+                    "unoTest",
+                    "uno-test@email.com",
+                    "uno-test",
+                    "test memo"
+            );
         }
     }
 
-
-    private ArticleCommentDto createArticleCommentDto(String content) {
-        return ArticleCommentDto.of(
-                1L,
-                1L,
-                createUserAccountDto(),
-                null,
-                content,
-                LocalDateTime.now(),
-                "Uno",
-                LocalDateTime.now(),
-                "Uno"
-        );
-    }
-
-    private UserAccountDto createUserAccountDto() {
-        return UserAccountDto.of(
-                "unoTest",
-                "uno-test@email.com",
-                "uno-test",
-                "test memo"
-        );
-    }
 }
